+++ NEWS.md
@@ -29,6 +29,7 @@
   - `chi.squared` -> `FSelector_chi.squared`
   - `relief` -> `FSelector_relief`
   - `oneR` -> `FSelector_oneR`
+* filter `mrmr` was removed because package _mRMRe_ was removed from CRAN on 11-01-2019 (@pat-s, #2532).
 * fixed a bug related to the loading of namespaces for required filter packages (@pat-s, #2483)
   
 ## learners - new